+++ banhammer/__init__.py
@@ -8,9 +8,4 @@ from . import reaction as ReactionHelper
 from . import reddithelper as RedditHelper
 from .subreddit import Subreddit
 from . import yaml as YAMLParser
-
-__title__ = "banhammer"
-__author__ = "Mariavi"
-__license__ = "GPL3.0"
-__copyright__ = "Copyright 2019 Mariavi"
-__version__ = "1.14.0"
+from . import config as Config