+++ banhammer/config.py
@@ -1,0 +1,11 @@
+__title__ = "banhammer"
+__author__ = "Mariavi"
+__license__ = "GPL v3.0"
+__copyright__ = "© Copyright 2019 Mariavi"
+__version__ = "1.17.0"
+
+BOT_VERSION_TEXT = "Banhammer.py v" + __version__
+
+BOT_FOOTER = "^({} | /r/TheBanhammerBot | Join us on) ^[Discord](https://discord.gg/9JrGC8f)".format(BOT_VERSION_TEXT)
+BOT_DISCLAIMER = "*This action was performed by the users of the [Banhammer.py](https://www.github.com/Dan6erbond/Banhammer.py) framework. " \
+                 "Please [contact the moderators of this subreddit]({}) if you have any questions or concerns.*\n\n" + BOT_FOOTER